#48: Add admin module/cog, move some commands there and add a permissions check
