+++ bot/admin/__init__.py
@@ -1,0 +1,2 @@
+"""Init file for making modules available outside of this package."""
+
